refactor(router): split app_router.dart by feature module (#4508)
app_router.dart had grown to 1,471 lines with 62 commits in 90 days,
making it a cross-feature merge hotspot (epic #4339 Wave 2). It declared
all 86 routes inside one Provider<GoRouter> closure.

Split the route declarations into per-feature modules under
lib/router/routes/ (shell, video, search, messaging, minor-account-review,
lists, auth, apps, settings, profile, library). app_router.dart is now a
109-line composition root that only builds the GoRouter and spreads the
module functions. Each module exposes a List<RouteBase> function; only
lists and apps take a Ref (they read providers in a route redirect/builder).

The redirect closure plus its auth/minor-review helpers and the
library-private navigation flags (_hasNavigated, _suppressNext…) move to
app_router_redirect.dart as `part of` app_router.dart, so those flags stay
shared with the public setters and every exported symbol keeps its import
path — callers and tests are unchanged (compat re-export mirrors #4506's
app_providers split). router_guards.dart holds the two pure helpers shared
by a route module and a direct-importing test, avoiding an import cycle.

No route path or name changes: the external URL contract (universal links,
OAuth deep links, signer callback) and go_router's global name-uniqueness
are preserved. Retargeted the two static source-introspection guards to
their new files and added a people-lists route-order guard
(/people-lists/new must precede /people-lists/:listId).
