+++ mobile/scripts/baseline/file_sizes.txt
@@ -11,7 +11,6 @@ lib/main.dart	2715
 lib/providers/clip_manager_provider.dart	894
 lib/providers/individual_video_providers.dart	1410
 lib/providers/video_editor_provider.dart	1051
-lib/router/app_router.dart	1562
 lib/screens/auth/invite_gate_screen.dart	819
 lib/screens/auth/nostr_connect_screen.dart	817
 lib/screens/creator_analytics_screen.dart	1406
